refactor: moved showDebug, showCircles, and circleType out of information and into controls slice
