+++ src/redux/informationSlice.ts
@@ -1,14 +1,6 @@
 import { createSlice } from "@reduxjs/toolkit";
 
-export enum CircleType {
-    INCREMENTAL = "incremental",
-    SOLAR = "solar",
-}
-
 export interface InformationState {
-    showDebug: boolean;
-    showCircles: boolean;
-    circleType: CircleType;
     // Universe debug variables
     numActiveBodies: number;
     numStars: number;
@@ -23,9 +15,6 @@ export interface InformationState {
 }
 
 const initialState: InformationState = {
-    showDebug: false,
-    showCircles: false,
-    circleType: CircleType.INCREMENTAL,
     numActiveBodies: 0,
     numStars: 0,
     maxVertexUniformVectors: 0,
@@ -41,15 +30,6 @@ export const informationSlice = createSlice({
     name: "information",
     initialState,
     reducers: {
-        toggleDebug: (state) => {
-            state.showDebug = !state.showDebug;
-        },
-        toggleCircles: (state) => {
-            state.showCircles = !state.showCircles;
-        },
-        toggleCircleType: (state) => {
-            state.circleType = state.circleType === CircleType.INCREMENTAL ? CircleType.SOLAR : CircleType.INCREMENTAL;
-        },
         setNumActiveBodies: (state, action) => {
             state.numActiveBodies = action.payload;
         },
